+++ go-cosmwasm/api/lib_mock.go
@@ -80,7 +80,7 @@ func ReleaseCache(cache Cache) {
 	// C.release_cache(cache.ptr)
 }
 
-func InitEnclaveRuntime(ModuleCacheSize uint8) error {
+func InitEnclaveRuntime(ModuleCacheSize uint16) error {
 	return nil
 }
